Tarea/requirements.txt
asgiref==3.2.7
astroid==1.5.3
autopep8==1.3.2
bleach==2.0.0
BTrees==4.6.1
certifi==2018.10.15
cffi==1.13.2
chardet==3.0.4
cycler==0.10.0
decorator==4.1.2
dj-database-url==0.5.0
Django==3.0.5
django-cors-headers==3.2.1
django-heroku==0.3.1
djangorestframework==3.11.0
entrypoints==0.2.3
flake8==3.4.1
gunicorn==20.0.4
html5lib==0.999999999
idna==2.7
img2pdf==0.3.3
ipykernel==4.6.1
ipython==6.1.0
ipython-genutils==0.2.0
ipython-sql==0.3.9
ipywidgets==6.0.0
isort==4.2.15
jedi==0.10.2
Jinja2==2.9.6
jsonschema==2.6.0
jupyter==1.0.0
jupyter-client==5.1.0
jupyter-console==5.1.0
jupyter-core==4.3.0
kiwisolver==1.0.1
lazy-object-proxy==1.3.1
MarkupSafe==1.0
matplotlib==2.2.0
mccabe==0.6.1
mistune==0.7.4
nbconvert==5.2.1
nbformat==4.3.0
notebook==5.0.0
numpy==1.14.1
pandocfilters==1.4.2
pep8==1.7.0
persistent==4.5.1
pexpect==4.2.1
pickleshare==0.7.4
Pillow==7.0.0
pipenv==2018.11.26
prettytable==0.7.2
prompt-toolkit==1.0.15
psycopg2==2.7.5
ptyprocess==0.5.2
pycodestyle==2.3.1
pycparser==2.19
pyflakes==1.5.0
Pygments==2.2.0
pylint==1.7.2
pymongo==3.9.0
pyparsing==2.2.0
PyQt5==5.9
PyRECONSTRUCT==2.2.0
python-dateutil==2.6.1
pytz==2018.3
pyzmq==16.0.2
qtconsole==4.3.0
requests==2.20.1
scribd-dl==0.4.5
selenium==3.141.0
simplegeneric==0.8.1
simpy==3.0.10
sip==4.19.3
six==1.11.0
SQLAlchemy==1.3.7
sqlparse==0.3.0
terminado==0.6
testpath==0.3.1
tornado==4.5.1
traitlets==4.3.2
unicorn==1.0.2rc2
urllib3==1.24.1
virtualenv==16.7.8
virtualenv-clone==0.5.3
wcwidth==0.1.7
webencodings==0.5.1
whitenoise==5.0.1
widgetsnbextension==2.0.0
wrapt==1.10.11
zope.interface==4.7.1